feat(ui): add Tatoeba tab to the typing-test mode modal
Extract the MonkeyType tab's pack list + version-check/update banner into
a reusable, provider-parameterized LanguagePackTab, and add a third
Tatoeba tab that lists Hub-distributed sentence packs the same way.
Picking a pack switches the config to the new tatoeba mode; picking a
MonkeyType language leaves tatoeba (restoring the last words/time/quote
config, as fileImport already does). The tatoeba mode hides the
pattern/units settings bar and shows the pack language as the mode label.

Only one tab is mounted at a time, so the pack rows/search/update-banner
keep identical test ids across providers. Adds tabTatoeba i18n (en/ja)
and tests for the new tab.
